% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R, R/dustmicrobiome.R
\docType{data}
\name{data-fungi}
\alias{data-fungi}
\alias{fungi}
\title{Dust Microbiome OTU and Sample Data for ITS (fungal) amplicons.}
\format{a phyloseq object with three slots
\describe{
\item{\code{sample_data}}{A dataframe with Sample information}
\item{\code{otu_table}}{A matrix with  OTU information. Rows are OTUs,
                        columns are samples, the value is the observer
                        OTU-per-column count}
\item{\code{tax_table}}{A matrix containing taxonomy information.
                        Rows are OTUs, columns are units of taxonomy.}
}}
\source{
http://figshare.com/articles/1000homes/1270900

http://figshare.com/articles/1000homes/1270900
}
\description{
Phyloseq object containing OTU frequencies of ITS amplicon data..

Phyloseq object containing OTU frequencies of ITS amplicon data..
}
\seealso{
\code{\link[phyloseq]{phyloseq}}
  \code{\link[phyloseq]{sample_data}}
  \code{\link[phyloseq]{otu_table}}
  \code{\link[phyloseq]{tax_table}}

\code{\link[phyloseq]{phyloseq}}
  \code{\link[phyloseq]{sample_data}}
  \code{\link[phyloseq]{otu_table}}
  \code{\link[phyloseq]{tax_table}}
}
\keyword{data}

