% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dustmicrobiome.R
\docType{package}
\name{dustmicrobiome}
\alias{dustmicrobiome}
\alias{dustmicrobiome-package}
\title{Dust Microbiome Data.}
\source{
http://figshare.com/articles/1000homes/1270900
}
\description{
This package is the OTU data from the paper "Continental-scale distributions
of dust-associated bacteria and fungi" from Noah Fierer's group. The data has
been made available on \url{http://figshare.com/articles/1000homes/1270900}{figshare}.
This package proveds that data as a set of ready-to-analyze.
}
\author{
Zach Charlop-Powers \email{zcharlop@rockefeller.edu}
}
\references{
\href{http://intl.pnas.org/content/112/18/5756.abstract}{Continental-scale
                  distributions of dust-associated bacteria and fungi}
}
\keyword{package}

